Simplify typescript-fetch types
