+++ .npmrc
@@ -1,7 +1,9 @@
-use-node-version=20.8.1
+use-node-version=20.17.0
 save-workspace-protocol=rolling
 child-concurrency=1
 link-workspace-packages=deep
 prefer-workspace-packages=true
 disallow-workspace-cycles=true
 shamefully-hoist=true
+manage-package-manager-versions=true
+engine-strict=true
